+++ example/click_functions_helper_example.dart
@@ -3,4 +3,5 @@ import 'package:click_functions_helper/click_functions_helper.dart';
 void main() {
   print('onlyNumbers: ${FunctionsHelper.onlyNumbers('1a2b3c/+')}');
   print('appPath: ${FunctionsHelper.appPath('abc.txt')}');
+  print('cript:teste ${FunctionsHelper.encriptar('teste')}');
 }