+++ CHANGELOG.md
@@ -1,4 +1,18 @@
-## 1.0.2
+## 1.0.10
+- image: ^4.0.15
+## 1.0.9
+- crypto
+## 1.0.8
+- adicionado funções isValidCPF, isValidCNPJ e isValidCPForCNPJ
+## 1.0.7
+- testando função para converter image to base64
+## 1.0.6
+- testando função para converter image to base64
+## 1.0.5
+- testando função para converter imageProvider to base64
+## 1.0.4
+- Adicionando a função para converter image to base64
+## 1.0.3
 - Adicionado a função convert base64 to image.
 ## 1.0.2
 - Adicionado a função getCRC16.